requirement.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.14.16
botocore==1.17.16
certifi==2020.6.20
cffi==1.14.0
channels==2.3.0
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==2.2.4
django-filter==2.3.0
django-rest-swagger==2.2.0
django-s3direct==1.1.7
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
hiredis==1.1.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==0.5.6
msgpack-python==0.5.6
ntp==1.0.0a1
openapi-codec==1.3.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
razorpay==1.2.0
redis==2.10.6
requests==2.24.0
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
twilio==6.43.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
zope.interface==5.1.0
